+++ CHANGELOG.md
@@ -12,3 +12,4 @@ HEAD
 * Fix PubSubclient callback and include onDataReceived
 * Integrate addTopicSub method for esp8266. Handle connection state and topic subscription on connection
 * Add client buffer set/get for esp8266 target
+* Add publish method with topic, data and data length